Slogging through the display classes, attempt to make them as quick to execute and as low on memory consumption as I can.
So far the classes created use a tree system for culling out objects not to be rendered, and only iterate of the objects to be displayed once, which should ensure quick execution. The culling operation relies on Rectangle's intersect() method, which should be efficient enough for all circumstances. There are quite a few instances of AffineTransform produced as a result of rendering: They're used to position images on the screen and were chosen due to their ability to be composited (good for having children displayed relative to their parents) and the ease with which they can be applied to both BufferedImages and Shapes. I believe (though I have no evidence to back it up) that AffineTransforms are relatively lightweight, so that this shouldn't be a problem.