Changing the architecture once again. Changes aren't complete but here's a summary of the new architecture:
Game responsiblity is divided into collision, culling, display, logic, and organization.
Each aspect has an interface, with a default implementation that has bridge methods to another interface that can be instantiated and wrapped by the default implementation.
There is a heirarchy of responsiblity for the default implementations: Cullable->Node->Displayable->Stated->Collidable.
Instead of compositing each interface for each implementation (which leads to complexity with the factories), each step up the chain takes the interface at the current level and adds it to a composite interface for the previous level: Cullable->Node->Node + Displayable = Sprite->Sprite + Stated = Entity->Entity + Collidable = Solid.
Finally, the default implementations (which are instances of composite interfaces) are created via abstract factories, which take in constructors, the factory for the preivous level, and have an abstract method that generates the implementation for the default instance.